pcubes_ws/requirements.txt
django
pm4py
djangorestframework
markdown
django-filter
pymongo